requirements.txt
pytorch
albumentations
tqdm
torchinfo
numpy
PIL
cv2